Update build configuration and documentation (#7)
- Added build_log.txt to .gitignore to prevent unnecessary log files from being tracked.
- Updated BUILD.md to include new toolchain notes and build instructions for C23 compliance.
- Enhanced CMakeLists.txt to enforce C23 standards and added options for stricter warning settings.
- Improved README.md with new standards section detailing C23 target and internal code preferences.
- Made minor adjustments to various source files to suppress unused parameter warnings and improve code clarity.

+++ src/qcommon/cm_load.c
@@ -653,14 +653,14 @@ void CM_LoadMap( const char *name, qboolean clientload, int *checksum ) {
 	if ( !buf ) {
 		Com_Error( ERR_DROP, "%s: couldn't load %s", __func__, name );
 	}
-	if ( length < sizeof( dheader_t ) ) {
+	if ( (size_t) length < sizeof( dheader_t ) ) {
 		Com_Error( ERR_DROP, "%s: %s has truncated header", __func__, name );
 	}
 
 	*checksum = cm.checksum = LittleLong( Com_BlockChecksum( buf, length ) );
 
 	header = *(dheader_t *)buf;
-	for ( i = 0; i < sizeof( dheader_t ) / sizeof( int32_t ); i++ ) {
+	for ( i = 0; (size_t) i < sizeof( dheader_t ) / sizeof( int32_t ); i++ ) {
 		( (int32_t *)&header )[i] = LittleLong( ( (int32_t *)&header )[i] );
 	}
